Allow copy MultiDataEntry from view
